requirements.txt
asgiref==3.8.1
coverage==7.6.8
Django==5.1.3
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.2
